+++ include/libfly/system/SoA.hpp
@@ -42,16 +42,6 @@ namespace fly::system {
     template <typename... Pr, typename... Mx>
     struct different_SoA<SoA<Pr...>, SoA<Mx...>> : std::true_type {};
 
-    //
-    template <typename...>
-    struct find : std::false_type {};
-
-    template <typename T, typename... Ts>
-    struct find<T, T, Ts...> : std::true_type {};
-
-    template <typename T, typename U, typename... Ts>
-    struct find<T, U, Ts...> : find<T, Ts...> {};
-
   }  // namespace detail
 
   /**
@@ -81,12 +71,12 @@ namespace fly::system {
   class SoA : public detail::Adaptor<Pr>... {
   public:
     /**
-     * @brief True if this SoA is a pure view i.e. all its properties  are reference types.
+     * @brief True if this SoA is a pure view i.e. all its properties are reference types.
      */
     static constexpr bool owns_none = (std::is_reference_v<Pr> && ...);
 
     /**
-     * @brief True if this SoA is purely owning i.e. non of its properties  are reference types.
+     * @brief True if this SoA is purely owning i.e. non of its properties are reference types.
      */
     static constexpr bool owns_all = (!std::is_reference_v<Pr> && ...);
 
@@ -131,7 +121,7 @@ namespace fly::system {
      * constructable from Other's adaptors.
      */
     template <typename Other>
-    static constexpr bool constructible = different_SoA_v<Other> && (std::is_constructible_v<detail::Adaptor<Pr>, Other&&> && ...);
+    static constexpr bool constructible = different_SoA_v<Other> && (std::is_constructible_v<detail::Adaptor<Pr>, Other> && ...);
 
   public:
     /**
@@ -140,8 +130,26 @@ namespace fly::system {
      * Only SFINE enabled if this SoA owns non of its arrays and this SoA is actually constructable from other.
      *
      */
-    template <typename T, typename = std::enable_if_t<owns_none && constructible<T>>>
-    SoA(T&& other) : detail::Adaptor<Pr>(std::forward<T>(other))..., m_size(other.size()) {
+    template <typename... T, typename = std::enable_if_t<owns_none && constructible<SoA<T...> const&>>>
+    SoA(SoA<T...> const& other) : detail::Adaptor<Pr>(other)..., m_size(other.size()) {}
+
+    /**
+     * @brief Implicitly construct a new SoA object from SoA 'other' with different properties .
+     *
+     * Only SFINE enabled if this SoA owns non of its arrays and this SoA is actually constructable from other.
+     *
+     */
+    template <typename... T, typename = std::enable_if_t<owns_none && constructible<SoA<T...>&>>>
+    SoA(SoA<T...>& other) : detail::Adaptor<Pr>(other)..., m_size(other.size()) {}
+
+    /**
+     * @brief Implicitly construct a new SoA object from SoA 'other' with different properties .
+     *
+     * Only SFINE enabled if this SoA owns non of its arrays and this SoA is actually constructable from other.
+     *
+     */
+    template <typename... T, typename = std::enable_if_t<owns_none && constructible<SoA<T...>&&>>>
+    SoA(SoA<T...>&& other) : detail::Adaptor<Pr>(std::move(other))..., m_size(other.size()) {
       // OK to ``std::move`` ``other`` multiple times as the detail::Adaptor constructor will only move its
       // corresponding base slice.
     }
@@ -152,8 +160,30 @@ namespace fly::system {
      * SFINE enabled if this SoA owns some of its arrays and this SoA is actually constructable from other.
      *
      */
-    template <typename T, typename = std::enable_if_t<!owns_none && constructible<T>>, typename = void>
-    explicit SoA(T&& other) : detail::Adaptor<Pr>(std::forward<T>(other))..., m_size(other.size()) {
+    template <typename... T, typename = std::enable_if_t<!owns_none && constructible<SoA<T...> const&>>, typename = void>
+    explicit SoA(SoA<T...> const& other) : detail::Adaptor<Pr>(other)..., m_size(other.size()) {
+      // The unnamed template parameter is a dummy to distinguish this from the implicit version.
+    }
+
+    /**
+     * @brief Explicitly construct a new SoA object from SoA 'other' with different properties .
+     *
+     * SFINE enabled if this SoA owns some of its arrays and this SoA is actually constructable from other.
+     *
+     */
+    template <typename... T, typename = std::enable_if_t<!owns_none && constructible<SoA<T...>&>>, typename = void>
+    explicit SoA(SoA<T...>& other) : detail::Adaptor<Pr>(other)..., m_size(other.size()) {
+      // The unnamed template parameter is a dummy to distinguish this from the implicit version.
+    }
+
+    /**
+     * @brief Explicitly construct a new SoA object from SoA 'other' with different properties .
+     *
+     * SFINE enabled if this SoA owns some of its arrays and this SoA is actually constructable from other.
+     *
+     */
+    template <typename... T, typename = std::enable_if_t<!owns_none && constructible<SoA<T...>&&>>, typename = void>
+    explicit SoA(SoA<T...>&& other) : detail::Adaptor<Pr>(std::move(other))..., m_size(other.size()) {
       // OK to ``std::move`` ``other`` multiple times as the detail::Adaptor constructor will only move its
       // corresponding base slice. The unnamed template parameter is a dummy to distinguish this from the implicit version.
     }