+++ include/libfly/system/typemap.hpp
@@ -105,8 +105,9 @@ namespace fly::system {
     /**
      * @brief Construct a new TypeMap by slicing a different kind of TypeMap.
      */
-    template <typename... U, typename = std::enable_if_t<!detail::same_properties<TypeMap, U...>::value>>
-    explicit TypeMap(TypeMap<U...> map) : SOA(static_cast<SoA<Type, U...>>(std::move(map))) {}
+    template <typename... U, typename = std::enable_if_t<!detail::same_properties<TypeMap, U...>::value
+                                                         && std::is_constructible_v<SOA, SoA<Type, U...> const&>>>
+    explicit TypeMap(TypeMap<U...> const& map) : SOA(static_cast<SoA<Type, U...> const&>(map)) {}
 
     /**
      * @brief Fetch the number of types stored in the TypeMap